requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
pandas==2.1.4
openpyxl==3.1.2
selenium==4.16.0
webdriver-manager==4.0.1
Werkzeug==3.0.1